requirements.txt
python-dotenv
Pillow
opencv-python
pyzbar
requests
keyboard
python-barcode
pytest
reportlab
